Provide `TryFrom<&[T]>` implementation for `<&[[T; N]]>`
The `TryFrom<&[T]>` implementation for `<&[[T; N]]>` allows converting
a slice into a slice of N-element chunks while asserting that the
slice is evenly divided into those chunks.  This is `[T]::as_chunks`
but makes it more concise to check that check that there's no
reminder.  For example, something like:

    let (chunks, remainder) = slice.as_chunks::<N>();
    if !remainder.is_empty() {
        return Err(SomeError);
    }

becomes:

    let chunks = <&[[T; N]]>::try_from(slice).map_err(|_| SomeError)?;