core: implement Pattern<&[T]>; get rid of SlicePattern
Implement Haystack<&[T]> and corresponding Pattern<&[T]> for &T and
&[T].  That is, provide implementation for searching for elements and
subslices in slices.  This replaces SlicePattern type.

To make use of those new implementations, provide a few new methods on
[T] type modelling them after types on str.  Specifically, introduce
{starts,ends}_with_pattern, find, rfind, {split,rsplit}_once and
trim{,_start,_end}_matches.

Note that starts_with and ends_with methods were already present.  This
changes them to be generic on the type of pattern.  This is technically
API breaking change (as it prevents type inference and autoderef) alas
I believe it might be in spriti of RFC 1105¹ which states:

> Minor changes should require at most minor amounts of work upon
> upgrade.  For example, changes that may require occasional type
> annotations or use of UFCS to disambiguate are not automatically
> ‘major’ changes. (But in such cases, one must evaluate how widespread
> these ‘minor’ changes are).

Alternative approach would be to use starts_with_pattern and
ends_with_pattern names instead.

For the time being, this change doesn’t implement functions returning
iterators such as split or matches which in str type are built on top
of the Pattern API.

¹ https://rust-lang.github.io/rfcs/1105-api-evolution.html