core: add core::pattern::Predicate wrapper type
To work around orphan rules, introduce a wrapper type for predicate
functions to be used as pattern.  The wrapper is necessary because
otherwise we end up with conflicting implementations if we want `&T`
and `&FnMut(T) -> bool` patterns over a slice.  In particular, see:

    error[E0119]: conflicting implementations of trait `pattern::Pattern<&[_]>` for type `&_`

    | impl<'hs, 'p, T: PartialEq> Pattern<&'hs [T]> for &'p T {
    | ------------------------------------------------------- first implementation here

    | impl<'hs, T, F: FnMut(&'hs T) -> bool> Pattern<&'hs [T]> for F {
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ conflicting implementation for `&_`

With this new wrapper, it’s now possible to have an implementation for
`&T` and another for a `Predicate`, i.e.:

    impl<'hs, T: PartialEq> Pattern<&'hs [T]> for &T { /* ... */ }
    impl<'hs, T, F: FnMut(T) -> bool> Pattern<&'hs [T]> for Predicate<F> { /* ... */ }

Technically this can be done with use of copious amounts of nightly
features (namely auto_traits and negative_impls).  However, for the
time being leave it with a wrapper for somewhat more verbose but
hopefully conceptually simpler implementation.